refactor(web): dedup sidebar primitives flagged by /simplify
Post-release cleanup over the v0.5.0..HEAD Signal diff (#311). Extracts
the sidebar's module-private primitives so the surviving hand-rolled
copies collapse onto them:

- useOutsideDismiss / useDismissableMenu → hooks/use-outside-dismiss.ts;
  page-menu.tsx now consumes the shared hook (also gains Escape-to-close)
  instead of its own document mousedown listener.
- navRowClass / flyoutRowClass → components/sidebar/row-styles.ts;
  admin/layout.tsx settings nav consumes navRowClass() rather than a
  copy-pasted class string (CLAUDE.md's "one shared navRowClass").
- Add loadString/saveString guards; the current-space localStorage
  read/write share one try/catch instead of two inline blocks.
- Remove dead --ease-emphasis token (byte-identical unused duplicate of
  --ease-signature; contradicts #310's "one curve").

Pure extraction/rewiring — no behavior change beyond page-menu Escape.
Token guard, eslint, and tsc --noEmit all green.
